requirements.txt
streamlit
tensorflow-cpu==2.9.0
numpy==1.26.4
pandas==2.3.3
scikit-learn==1.7.2
opencv-python-headless==4.12.0
matplotlib==3.10.8
Pillow==12.0.0